api/requirements.txt
fastapi==0.115.6
tortoise-orm==0.24.0
httpx==0.28.1
aiomysql==0.2.0
python-dotenv==1.0.1
python-jose==3.4.0
passlib==1.7.4
python-multipart==0.0.20